Split componentFactory factory into own file.
