+++ src/componentFactory.js
@@ -1,34 +1,8 @@
 import React from 'react'
-import makeGetStyleForProps from './makeGetStyleForStates'
+import factory from './factory'
 
 const identity = _ => _
 
-export const instyledWithTransform = inputTransform =>
-  (styleDefinitions, { component = 'div', name = 'instyled', staticProps = {} } = {}) => {
-    const getStyleForProps =
-      makeGetStyleForProps(inputTransform(styleDefinitions))
-
-    const Wrapper = props => {
-      const { children, mergeStyle, ...etc } = props
-      let style = getStyleForProps(etc)
-
-      if (mergeStyle != null) {
-        style = Object.assign({}, style, mergeStyle)
-      }
-
-      return React.createElement(
-        component,
-        { ...staticProps, style, ...etc },
-        children
-      )
-    }
-
-    if (name != null && typeof name === 'string') {
-      name = name[0].toUpperCase() + name.slice(1)
-      Wrapper.displayName = name
-    }
-
-    return Wrapper
-  }
+export const instyledWithTransform = factory(React.createElement)
 
 export const instyled = instyledWithTransform(identity)
