+++ test/index.js
@@ -1,3 +1,4 @@
 require('babel-core/register')
 
 require('./test.js')
+require('./factory.js')
